# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Exam Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exams:
    get:
      operationId: listExams
      tags: [exams]
      summary: List exams
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
        - name: examType
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
    post:
      operationId: createExam
      tags: [exams]
      summary: Create exam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamInput"
      responses:
        "201":
          description: Created exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /exams/{id}:
    get:
      operationId: getExam
      tags: [exams]
      summary: Get exam by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamDetail"
    patch:
      operationId: updateExam
      tags: [exams]
      summary: Update exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamUpdate"
      responses:
        "200":
          description: Updated exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
    delete:
      operationId: deleteExam
      tags: [exams]
      summary: Delete exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /exams/{id}/publish:
    patch:
      operationId: toggleExamPublish
      tags: [exams]
      summary: Toggle exam publish status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"

  # SECTIONS
...
  /papers:
    get:
      operationId: listPapers
      tags: [papers]
      summary: List previous year papers
      parameters:
        - name: examType
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of papers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Paper"
    post:
      operationId: createPaper
      tags: [papers]
      summary: Upload previous year paper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperInput"
      responses:
        "201":
          description: Created paper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paper"
  /papers/{id}:
    get:
      operationId: getPaper
      tags: [papers]
      summary: Get paper by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Paper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paper"
    delete:
      operationId: deletePaper
      tags: [papers]
      summary: Delete paper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # VIOLATIONS
...
  /feeds:
    get:
      operationId: listFeeds
      tags: [feeds]
      summary: List feed posts
      parameters:
        - name: published
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of feeds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Feed"
    post:
      operationId: createFeed
      tags: [feeds]
      summary: Create feed post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedInput"
      responses:
        "201":
          description: Created feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
  /feeds/{id}:
    get:
      operationId: getFeed
      tags: [feeds]
      summary: Get feed by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
    patch:
      operationId: updateFeed
      tags: [feeds]
      summary: Update feed post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedUpdate"
      responses:
        "200":
          description: Updated feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
    delete:
      operationId: deleteFeed
      tags: [feeds]
      summary: Delete feed post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # DASHBOARD
...
components:
  schemas:
...
    Exam:
      type: object
      required: [id, title, examType, duration, totalMarks, active, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        examType:
          type: string
        duration:
          type: integer
        totalMarks:
          type: integer
        negativeMarks:
          type: number
        instructions:
          type: ["string", "null"]
        active:
          type: boolean
        hasSectionalTimer:
          type: boolean
        startDate:
          type: ["string", "null"]
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        questionCount:
          type: ["integer", "null"]
        attemptCount:
          type: ["integer", "null"]
    ExamDetail:
      type: object
      required: [id, title, examType, duration, totalMarks, active, createdAt, sections, questions]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        examType:
          type: string
        duration:
          type: integer
        totalMarks:
          type: integer
        negativeMarks:
          type: number
        instructions:
          type: ["string", "null"]
        active:
          type: boolean
        hasSectionalTimer:
          type: boolean
        startDate:
          type: ["string", "null"]
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        sections:
          type: array
          items:
            $ref: "#/components/schemas/Section"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
    ExamInput:
      type: object
      required: [title, examType, duration, totalMarks]
      properties:
        title:
          type: string
        description:
          type: string
        examType:
          type: string
        duration:
          type: integer
        totalMarks:
          type: integer
        negativeMarks:
          type: number
        instructions:
          type: string
        active:
          type: boolean
        hasSectionalTimer:
          type: boolean
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
    ExamUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        examType:
          type: string
        duration:
          type: integer
        totalMarks:
          type: integer
        negativeMarks:
          type: number
        instructions:
          type: string
        active:
          type: boolean
        hasSectionalTimer:
          type: boolean
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time

    # SECTIONS
    Section:
      type: object
      required: [id, name, examId, duration, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        examId:
          type: integer
        duration:
          type: integer
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Question:
      type: object
      required: [id, questionText, optionA, optionB, optionC, optionD, correctAnswer, marks, createdAt]
      properties:
        id:
          type: integer
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
        explanation:
          type: ["string", "null"]
        marks:
          type: number
        negativeMarks:
          type: number
        difficulty:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        examId:
          type: ["integer", "null"]
        sectionId:
          type: ["integer", "null"]
        subjectId:
          type: ["integer", "null"]
        topicId:
          type: ["integer", "null"]
        quizId:
          type: ["integer", "null"]
        topicMockId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    Paper:
      type: object
      required: [id, title, examType, fileUrl, fileType, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        examType:
          type: string
        year:
          type: ["integer", "null"]
        fileUrl:
          type: string
        fileType:
          type: string
        fileName:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PaperInput:
      type: object
      required: [title, examType, fileUrl, fileType, fileName]
      properties:
        title:
          type: string
        examType:
          type: string
        year:
          type: integer
        fileUrl:
          type: string
        fileType:
          type: string
        fileName:
          type: string
        description:
          type: string

    # FEEDS
    Feed:
      type: object
      required: [id, title, description, isPublished, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        isPublished:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    FeedInput:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        isPublished:
          type: boolean
    FeedUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        isPublished:
          type: boolean

    # VIOLATIONS
...