# lib/api-zod/src/generated/types/examUpdate.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Api
 * AI Exam Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ExamUpdate {
  title?: string;
  description?: string;
  examType?: string;
  duration?: number;
  totalMarks?: number;
  negativeMarks?: number;
  instructions?: string;
  active?: boolean;
  hasSectionalTimer?: boolean;
  startDate?: Date;
  endDate?: Date;
}